+++ src/services/index.ts
@@ -2,6 +2,7 @@ export * from './frame.js';
 export * from './input.js';
 export * from './switch.js';
 export * from './speaker.js';
+export * from './volume.js';
 export * from './television.js';
 export * from './information.js';
 export * from './wrapper.js';