+++ src/types/device.ts
@@ -83,4 +83,5 @@ export type DeviceConfig = {
   device_id?: string;
 };
 
-export type DeviceOptions = 'Device.Disable' | 'Switch.DeviceName.Disable' | 'Frame.RealPowerMode' | 'Frame.ArtSwitch.Disable' | 'Frame.PowerSwitch.Disable';
+export type DeviceOptions =
+  'Device.Disable' | 'Volume.Slider.Enable' | 'Switch.DeviceName.Disable' | 'Frame.RealPowerMode' | 'Frame.ArtSwitch.Disable' | 'Frame.PowerSwitch.Disable';